[fix]: rename files to match styleguide standards
