+++ server/controllers/sendNotifications.js
